requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-text-splitters
chromadb
pypdf
pymupdf
rapidocr-onnxruntime
onnxruntime-gpu
python-docx
openpyxl
tqdm
python-dotenv
openai
chainlit
edge-tts
torch
torchvision
torchaudio